feat: support creating plan-only runs (#465)
Adds support for the specifying the `plan-only` attribute when creating
a run via the API:


https://developer.hashicorp.com/terraform/cloud-docs/api-docs/run#create-a-run

This overrides the `speculative` attribute when specified on the
configuration version for the run.

Note: throughout TFC documentation, both "speculative" and "plan-only"
are used to refer to the same thing: a run that only performs a plan and
does not follow through to an apply. I think Hashicorp are increasingly
using the latter term, "plan-only", because it is more descriptive. I've
followed suit in this PR and updated the OTF codebase to use variants of
the latter term and minimise references to "speculative" runs.