feat: create run using config from repo (#466)
Neither OTF (nor TFC) currently permits, in a single API call, creating
a run using configuration from a VCS repo (assuming the run's workspace
is connected to a VCS repo). Instead the user/client has to manually
invoke several operations:

1. retrieve tarball of vcs repo contents
2. create config version and upload tarball to config version
3. create run, specifying aforementioned config version

Step 1 can be particularly onerous for the client, because connectivity
and credentials are needed to access the repo, as well as necessary
tooling, (curl, git, or equivalent libs in a programmatic context).

This PR introduces a "magic string", `__pull_vcs__`, that if specified
in place of the configuration version ID when creating the run,
instructs OTF to perform those operations on behalf of the client.

Using a magic string rather than introducing another API endpoint means
the existing APIs and SDKs can be used, including `go-tfe`.

Note: I've also taken this opportunity to refactor out the "run
starter", which has been folded into the "run factory".