fix: state output values are json (#477)
State output values were being treated like a string throughout the
codebase, but of course they can be one of number of different terraform
data types.

Terraform types translate to JSON types, so this PR changes the output
value to be a `json.RawMessage`. The PR would also migrate the postgres
type from string to JSONB, but the `go-tfe` integration tests we use
rely on whitespace being preserved when comparing for equality; I tried
using JSON postgres type but that broke the queries we use because that
type doesn't "have an equality operator". So we'll migrate it to BYTEA,
which'll do just fine.