fix: finish events refactor (#509)
The events refactor from several weeks ago, #458, introduced postgres
triggers as a means of publishing events. But it was never 100%
completed, leaving behind many `Publish()` calls that do nothing.

Furthermore, the refactor introduced a regression: the agent event
subscriber was not also refactored and therefore it no longer received
cancelation events. When a user canceled a run, the run status would be
set to canceled but the terraform process would not receive an interrupt
signal. This PR fixes this, and introduces an integration test to
prevent re-occurance of the regression.