feat!: adding team member creates user if they don't exist (#525)
This PR changes the behaviour of the add team member functionality, both
via the UI and the API: when adding a user to a team, if the user does
not exist the user is created.

This is a deviation from TFC's behaviour. But then OTF already differs
quite a bit from TFC when it comes to users: in OTF, users are not
"invited" and are instead created the first time they login. There is no
email functionality because that would add a further dependency (a mail
relay).