chore: remove redundant CreateRun magic string (#555)
@fsaintjacques A heads up - I recall removing the magic string behaviour
from your software, so I'm pretty sure this is now safe to remove from
OTF.

(If you recall, when creating a run, we wanted the run to automatically
fetch the config from the git repo, so we introduced a magic string to
invoke this behaviour. But then I realised you could simply leave out
the configuration version ID from CreateRun and that invokes the same
behaviour).