refactor: move to upstream API tests (#535)
OTF has for a long while relied on the API tests from a fork of the
`go-tfe` repo. The fork was necessary because there were certain
differences in the API responses compared with those returned by TFE/C.

But recently more effort has gone into eliminating these differences.
Therefore the fork is less necessary than it once was and more and more
tests have instead been sourced from the upstream `go-tfe` repo.

This PR completes that migration and removes the necessity for the fork
entirely.