fix: resubscribe subsystems when their subscription is terminated (#593)
Fixes #577 

The events broker can legitimately terminate the subscription of a
subscriber, e.g. when the subscriber's event queue is full up. Various
subsystems such as the scheduler would, in such a scenario, simply
shutdown. Which in the case of #577 would mean no more runs could be
started.

This PR changes the behaviour of the subsystems: when their subscription
is terminated, they instead return an error, which triggers the
subsystem to be restarted (using a backoff algorithm) rather than stop
completely.