fix: fixed bug where proxy was ignored (#609)
This PR does two main things, 

## http.Transport proxy
In areas where we manually construct an http.Transport unless we
explicitly provide a Proxy field of http.ProxyFromEnvironment, we are
explicitly configuring the http package to ignore the proxy environment
variables. This means that otf ignores all proxy settings.

## oidc.NewProvider Timeout
Changes the oidc_authenticator to ensure that it throws an error after
waiting for more than ten seconds to construct the oidc provider. This
"fix" is intended to keep the tool silently waiting when the proxy
settings are ommitted (or other network issues are blocking in the
NewProvider func)

---------

Co-authored-by: Louis Garman <louisgarman@gmail.com>