Bump default terraform version to 1.6.2 (#621)
Note: a regression appears to be present in 1.6.2 that was not present
in 1.5.x: if a `terraform plan` is a mid-flow and the user sends the
process a ctrl-c signal, the process exits without canceling the run on
TFC/OTF. However, this regression only appears to impact `terraform
plan` and not `terraform apply`.

Another note: one of the integration tests for the recently added github
app was improperly implemented and would transiently fail. In order to
get this PR through, I've fixed it here.