refactor: dedicated API (#625)
This PR introduces a dedicated API for OTF. Until now, OTF has
piggybacked onto the TFC API.

This new dedicated API is not intended for external consumption. It is
not documented and is liable to change without notice. Like the TFC API,
it uses the `json:api` encoding as well as vanilla `json`.

(This PR replaces the previous attempt, #463, which took a more
ambitious approach using an OpenAPI spec and generating go code from
that. That PR was only half complete and would have a taken a lot longer
to complete).

So now OTF provides two APIs:

* TFE API: this is Hashicorp's documented API for TFE/TFC, used by
`terraform`, `go-tfe`, and the `tfe` provider.
* OTF API: this is bespoke to OTF, used by `otf` (the CLI) and
`otf-agent` (the agent process) to talk to `otfd` (the server process).