Fix local execution mode (#627)
Fixes #626.

The TFC [State Versions
API](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/state-versions#create-a-state-version)
has drastically changed. Starting with v1.6.0, terraform started using
the updated API for workspaces using the remote/cloud backends, i.e. TFC
and OTF, but the changes only impact local execution mode usage. Hence
#626 reported seeing issues with v1.6.0 and OTF.

This PR incorporates the changes to the State Versions API. Essentially
instead of:

* Create State Version with metadata and the state itself

There is now _also_ the ability to do this in two separate API calls:

* Create State Version with only metadata
* Upload state to that state version

And with that, a state version now has a _status_, which can be one of:

* pending
* finalized
* discarded

The implemented behaviour follows as much as possible that documented
for the API (see link above).