fix: agent error reporting (#628)
Fixes #626.

#625 broke the OTF API endpoint for creating a state version, which is
used by the agent to create and upload state. Fortunately this
regression was only introduced recently and did not get released.

#625 introduced another regression to the OTF API endpoint responsible
for finishing a run phase, which again is used by the agent. The
regression in particular resulted in errors failing to be reported,
which meant the first bug didn't get captured by the integration tests.

This PR addresses these bugs and adds an integration test specifically
for run errors.