Implement TFE API for Team Tokens (#624)
Towards #580 

This is the initial API implementation for Team Tokens. The TFE
integration tests pass, but I don't think the token can actually be used
to do anything as yet.

I've added `Team` as a sort of half-concept for the RBAC. It's not a
full category in it's own right, but uses the `User` team memberships to
authorize. Not sure if that's desirable, but it seemed a minimal effort
fit as I understand how the RBAC works.

---------

Co-authored-by: Louis Garman <louisgarman@gmail.com>