refactor: tokens (#646)
Refactor of the `tokens` package. With this PR, the `tokens` package has
been slimmed down to focus on the generation and verification of tokens.
Implementation of particular tokens - organization, team, user tokens,
etc - has been moved to respective packages. Ultimately the package
dependency hierarchy is simpler, broader, and flatter (a hallmark of any
good Go project).