feat: add option to skip plans for pull requests (#714)
Currently OTF automatically triggers a speculative run i.e. a plan-only
run for pull request events, when a PR is opened, when commits are
pushed to the PR, etc.

This PR adds the option to disable this via the UI. It also adds the
option to do so via the API (most of the plumbing was already there for
the API for the purposes of satisfying the go-tfe integration tests, but
it didn't actually do anything...).