fix(allocator): job capacity exceeded on runners (#723)
Fixes a bug which allowed a runner's job capacity to be exceeded.

The allocator is responsible for allocated jobs to a runner. It should
respect the runner's capacity, i.e. the max number of jobs it is
permitted to run at once. It keeps a tally of the number of jobs
allocated to each runner. However, the allocator would also subscribe to
runner events and whenever an event arrived, it would overwrite its
local tally with the tally embedded in the event. That's problematic
because an event is something that happened in the past, and it can be
stale, containing a lower tally than that maintained locally. The
allocator would then wrongly allocate another job to the runner.

This fix instead maintains a local tally for each runner in the
allocator and ignores the tallies in events.