fix: add `--allowed-origins` parameter (#773)
Fixes: #772

This gives otfd users a way to specify a comma-separated list of
hostnames to validate websocket upgrade requests against.

Add a command-line parameter (`--allowed-origins`), add it to markdown
docs, pass it through, split by commas, put the hostnames in a map, give
`websocket.Upgrader` a callback to check for the Origin hostname's
presence in the map.

A possible alternative would be to just return true in all cases. That
would make this patch much simpler; effectively a one-liner.

---------

Co-authored-by: Louis Garman <louisgarman@gmail.com>