fix: "Save variable set" button on the variableset edit page (#778)
When editing a variable set, the form has the wrong action URL. Clicking
the "Save" button results in a 404 error.

The form tries to submit to `.../edit`, but that's a GET endpoint. The
POST endpoint is `.../update`. See the route definitions
[here](https://github.com/leg100/otf/blob/master/internal/variable/web.go#L88-L89).

This patch fixes the form to submit to the right place.

---------

Co-authored-by: Louis Garman <louisgarman@gmail.com>