refactor: make events more real time (#780)
Currently inserted/updated/deleted rows in postgres trigger events. The
events contain the table name, the action, and the primary key ID of the
changed row. An event "broker" then retrieves the resource corresponding
to the ID.

This PR changes the content of the event to contain not only the ID but
the entire row. The broker unmarshals the row into an resource event.

This means events are now the actual resource at the time the event
occured, rather than something retrieved some time afterwards.

It also simplifies the logic somewhat.