fix(ui): various VCS fixes (#797)
* Fixes off-center Gitlab icon
* Previously, if a run was created via the UI, and its workspace was
connected to a VCS repo, then the username of the user that created the
run was set to the user who committed the last commit on the main branch
of the repo. This PR changes that to set the username to the user who
created the run via the UI. Phew.
* Clicking on the personal access token link now opens the link in a new
tab.
* The personal access token link for Gitlab resulted in a 404; that's
been fixed.