fix: AWS dynamic provider credentials invalid apply absolute paths (#822)
This fixes an issue with aws dynamic provider credentials.

A new tmpdir is created for each operation, so the absolute path for
plan will be different than apply. The apply step uses the plan output,
which means it doesn't read the inputs (I think?). Which means it
wouldn't use the newly generated config/creds for the provider dynamic
credentials and would try to look in the old path (which no longer
existed).

The changing path also means that any files or values created during the
plan will run into issues if they contain the absolute path. For
example, the aws profile references another file by absolute path.